tests(pix-editor): replace unit test by integration test for grid cell
